refactor: Move filters from assertions package to new filters package
